Fix reviewer round 3: fix test naming, move parser to shared module
- Fix test function names: test_parse_* (was missing underscore after test)
- Move parse_agents_from_mng_output to config/data_types.py (shared module)
  so neither testing.py nor cli/update.py depends on the other
